+++ inst/julia/asR.jl
@@ -1,9 +1,6 @@
-asDouble(x :: Real) = Float64(x)
-asDouble(x :: AbstractString) = parse(Float64, x)
-asDouble(x :: Bool) = ifelse(x, 1.0, 0.0)
-asDouble{T<:Real}(x :: AbstractArray{T}) = AbstractArray{Float64}(x)
-asDouble{T<:AbstractString}(x :: AbstractArray{T}) = [asDouble(xx) for xx in x]
-asDouble(x :: AbstractArray{Bool}) = [asDouble(xx) for xx in x]
+asDouble(x) = float(x)
+asDouble(x :: Complex) = asDouble(real(x))
+asDouble(x :: AbstractArray{T}) where {T <: Complex} = [asDouble(xx) for xx in x]
 
 asCharacter(x) = string(x)