+++ inst/julia/dispatch.jl
@@ -11,3 +11,37 @@ cumprod1(x) = cumprod(x)
 cumprod1(x :: Number) = cumprod1([x])
 
 tanpi(x) = sinpi(x) / cospi(x)
+
+unlist(x) = vcat(x...)
+
+rep(x, times) = repmat(vcat(x), times)
+
+function assign!(x :: AbstractArray, value :: AbstractArray, i)
+    try
+        setindex!(x, value, i)
+    catch e
+        commontype = promote_type(eltype(x), eltype(value))
+        x = AbstractArray{commontype}(x)
+        setindex!(x, value, i)
+    end
+end
+
+function assign!(x :: AbstractArray, value, i...)
+    try
+        setindex!(x, value, i...)
+    catch e
+        commontype = promote_type(eltype(x), typeof(value))
+        x = AbstractArray{commontype}(x)
+        setindex!(x, value, i...)
+    end
+end
+
+## Array related dispatching methods
+
+isMatrix(x :: AbstractArray) = length(size(x)) == 2
+isMatrix(x) = false
+
+isArray(x :: AbstractArray) = true
+isArray(x) = false
+
+dim(x) = vcat(size(x)...)